requirements.txt
fastapi~=0.115.12
uvicorn
openai>=1.0.0
python-dotenv~=1.1.0
pydantic~=2.11.5

SQLAlchemy~=2.0.41
pytest~=8.3.5
requests~=2.32.3
chromadb~=1.0.13
pandas~=2.3.0